beta-algocode-parser/requirements.txt
schedule
pytz
telebot
selenium
tabulate
psycopg3
python-dotenv